Rename and restructure chain within package
