+++ lib/app/app_module.dart
@@ -2,8 +2,8 @@ import 'package:credible/app/pages/credentials/blocs/scan.dart';
 import 'package:credible/app/pages/credentials/blocs/wallet.dart';
 import 'package:credible/app/pages/credentials/module.dart';
 import 'package:credible/app/pages/credentials/repositories/credential.dart';
-import 'package:credible/app/pages/did/chain.dart';
-import 'package:credible/app/pages/did/display.dart';
+import 'package:credible/app/pages/chain/detail.dart';
+import 'package:credible/app/pages/did/detail.dart';
 import 'package:credible/app/pages/on_boarding/module.dart';
 import 'package:credible/app/pages/profile/module.dart';
 import 'package:credible/app/pages/qr_code/bloc/qrcode.dart';
