+++ lib/app/pages/did/detail.dart
